Cardiac output

Partial blood flow of common compartments.

cardiac_output(
  total = 6.3,
  prop_lung = 1,
  prop_vrg = 0.798,
  prop_mus = 0.157,
  prop_fat = 0.044
)

Arguments

total

numeric, total cardiac output in l/min.

prop_lung

numeric, proportion of cardiac output to the lung, should be 1.00 in general.

prop_vrg

numeric, proportion of cardiac output to the vessel rich compartment.

prop_mus

numeric, proportion of cardiac output to the musle compartment.

prop_fat

numeric, proportion of cardiac output to the fat compartment.

Value

a named numeric with blood flow values (in l/min) for lung, vrg (vessel rich group, visceral: brain, heart, kidney, ...), mus (muscle, lean tissue: muscle, skin, subcutaneous tissue, ...), and fat (fat, fatty tissue: yellow marrow, fat) for the given anaesthetic.

References

Table 3 in Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973). A simplified digital method for predicting anesthetic uptake and distribution. Computers in Biology and Medicine, 3(4), 385-395. doi:10.1016/0010-4825(73)90004-8