Conductance

Conductance is defined as ability of the fluid to deliver anaesthetic agents across a partial pressure gradient.

conductance(flow, partition_coefficient, tp_factor = stp_factor())

Arguments

flow

numeric, blood/air flow in l/min.

partition_coefficient

numeric, blood:gas or tissue:gas partition coefficient.

tp_factor

numeric, temperature/pressure factor.

Value

numeric, conductance of the tissue in l/(min*atm).

Details

From Cowles et al.: "In the case of blood perfusing a tissue, the conductance is equal to the rate of blood flow multiplied by the blood:gas partition coefficient and the factor, T0/P0Ti. In the case of the alveolar gas ventilating the lungs, the conductance is the rate of alveolar ventilation multiplied by the factor, T0/P0Ti. The gas:gas partition coefficient is equal to 1.0, by definition."

References

Eqn. 6 in Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973). A simplified digital method for predicting anesthetic uptake and distribution. Computers in Biology and Medicine, 3(4), 385-395. doi:10.1016/0010-4825(73)90004-8